Refactoring: moved the different warmup challenges into their respective packages - package names are challenge ids (1st three letters)
