+++ src/test/java/befaster/solutions/SUM/SumTest.java
@@ -1,4 +1,4 @@
-package befaster.solutions;
+package befaster.solutions.SUM;
 
 import org.junit.Test;
